+++ fixed-vector/Data/Vector/Fixed/Mutable.hs
@@ -1,9 +1,11 @@
 {-# LANGUAGE DataKinds             #-}
 {-# LANGUAGE FlexibleContexts      #-}
+{-# LANGUAGE MagicHash             #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE PolyKinds             #-}
 {-# LANGUAGE Rank2Types            #-}
 {-# LANGUAGE ScopedTypeVariables   #-}
+{-# LANGUAGE TypeApplications      #-}
 {-# LANGUAGE TypeFamilies          #-}
 {-# LANGUAGE TypeOperators         #-}
 -- |
@@ -13,7 +15,6 @@
 module Data.Vector.Fixed.Mutable (
     -- * Mutable vectors
     Arity
-  , arity
   , Mutable
   , DimM
   , MVector(..)
@@ -41,11 +42,12 @@ module Data.Vector.Fixed.Mutable (
 import Control.Applicative  (Const(..))
 import Control.Monad.ST
 import Control.Monad.Primitive
-import Data.Typeable  (Proxy(..))
 import Data.Kind      (Type)
-import GHC.TypeLits
-import Data.Vector.Fixed.Cont (Dim,PeanoNum(..),Peano,Arity,Fun(..),Vector(..),ContVec,arity,apply,accum,length)
 import Prelude hiding (read,length,replicate)
+import GHC.Exts (proxy#)
+
+import Data.Vector.Fixed.Cont (Dim,PeanoNum(..),Arity,ArityPeano(..),Fun(..),Vector(..),
+                               ContVec,apply,accum,length)
 
 
 ----------------------------------------------------------------
@@ -56,10 +58,10 @@ import Prelude hiding (read,length,replicate)
 type family Mutable (v :: Type -> Type) :: Type -> Type -> Type
 
 -- | Dimension for mutable vector.
-type family DimM (v :: Type -> Type -> Type) :: Nat
+type family DimM (v :: Type -> Type -> Type) :: PeanoNum
 
 -- | Type class for mutable vectors.
-class (Arity (DimM v)) => MVector v a where
+class (ArityPeano (DimM v)) => MVector v a where
   -- | Copy vector. The two vectors may not overlap. Since vectors'
   --   length is encoded in the type there is no need in runtime checks.
   copy :: PrimMonad m
@@ -81,8 +83,8 @@ class (Arity (DimM v)) => MVector v a where
 
 
 -- | Length of mutable vector. Function doesn't evaluate its argument.
-lengthM :: forall v s a. (Arity (DimM v)) => v s a -> Int
-lengthM _ = arity (Proxy :: Proxy (DimM v))
+lengthM :: forall v s a. (ArityPeano (DimM v)) => v s a -> Int
+lengthM _ = peanoToInt (proxy# @(DimM v))
 
 -- | Create copy of vector.
 --
@@ -222,22 +224,22 @@ thaw v = clone =<< unsafeThaw v
 ----------------------------------------------------------------
 
 -- | Generic inspect implementation for array-based vectors.
-inspectVec :: forall v a b. (Arity (Dim v), IVector v a) => v a -> Fun (Peano (Dim v)) a b -> b
+inspectVec :: forall v a b. (ArityPeano (Dim v), IVector v a) => v a -> Fun (Dim v) a b -> b
 {-# INLINE inspectVec #-}
 inspectVec v
   = inspect cv
   where
     cv :: ContVec (Dim v) a
     cv = apply (\(Const i) -> (unsafeIndex v i, Const (i+1)))
-               (Const 0 :: Const Int (Peano (Dim v)))
+               (Const 0 :: Const Int (Dim v))
 
 -- | Generic construct implementation for array-based vectors.
-constructVec :: forall v a. (Arity (Dim v), IVector v a) => Fun (Peano (Dim v)) a (v a)
+constructVec :: forall v a. (ArityPeano (Dim v), IVector v a) => Fun (Dim v) a (v a)
 {-# INLINE constructVec #-}
 constructVec =
   accum step
         (\(T_new _ st) -> runST $ unsafeFreeze =<< st :: v a)
-        (T_new 0 new :: T_new v a (Peano (Dim v)))
+        (T_new 0 new :: T_new v a (Dim v))
 
 data T_new v a n = T_new Int (forall s. ST s (Mutable v s a))